+++ packages/Button/src/LinkButton.js
@@ -10,28 +10,28 @@ const LinkPropTypes = {
   children: PropTypes.node.isRequired,
   href: PropTypes.string.isRequired,
   isDisabled: PropTypes.bool,
-  isOpenNewTab: PropTypes.bool,
   kind: PropTypes.oneOf(Button.Kinds.ALL),
   size: PropTypes.oneOf(ShirtSizes.DEFAULT),
+  shouldOpenNewTab: PropTypes.bool,
   suffixIcon: PropTypes.node,
 };
 
 const LinkDefaultProps = {
   a11yText: null,
   isDisabled: false,
-  isOpenNewTab: false,
   kind: Button.Kinds.LINK,
+  shouldOpenNewTab: false,
   size: ShirtSizes.MEDIUM,
   suffixIcon: <NewTabIcon />,
 };
 
-const isOpenNewTabProps = {
+const shouldOpenNewTabProps = {
   target: "_blank",
   rel: "noopener noreferrer",
 };
 
 const LinkButton = React.forwardRef(
-  ({ a11yText, children, href, isDisabled, isOpenNewTab, kind, size, suffixIcon, ...moreProps }, ref) => {
+  ({ a11yText, children, href, isDisabled, shouldOpenNewTab, kind, size, suffixIcon, ...moreProps }, ref) => {
     return (
       <a
         aria-label={a11yText}
@@ -40,13 +40,13 @@ const LinkButton = React.forwardRef(
         isDisabled={isDisabled}
         kind={kind}
         size={size}
-        {...(isOpenNewTab ? isOpenNewTabProps : {})}
+        {...(shouldOpenNewTab ? shouldOpenNewTabProps : {})}
         {...moreProps}
         as={isDisabled ? "span" : "a"}
         ref={ref}
       >
         {children}
-        {kind === Button.Kinds.LINK && isOpenNewTab ? suffixIcon : null}
+        {kind === Button.Kinds.LINK && shouldOpenNewTab ? suffixIcon : null}
       </a>
     );
   }